refactor: Move /level/local/source to /level/local/encoding
