+++ internal/pkg/service/stream/storage/level/local/encoding/count/counter_test.go
