+++ internal/pkg/service/stream/storage/level/local/encoding/size/size.go
