+++ internal/pkg/service/stream/storage/level/local/encoding/writechain/closer.go
